move duplicate code for markers into shared. remove unused classname
