+++ src/lib/markers/implementations/TodayMarker.js
@@ -1,27 +1,11 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import {
+  createMarkerStylesWithLeftOffset,
+  createDefaultRenderer
+} from './shared'
 
-const criticalStyles = {
-  position: 'absolute',
-  top: 0,
-  bottom: 0,
-  // FIXME: are these critical styles?
-  width: '2px',
-  backgroundColor: 'red'
-}
-
-// FIXME: this is used in all marker implementations
-// REVIEW: might want to memoize this as it creates a new object
-// in each render which is passed to React component
-const createMarkerStylesWithLeftOffset = leftOffset => ({
-  ...criticalStyles,
-  left: leftOffset
-})
-
-// eslint-disable-next-line
-const defaultRenderer = ({ styles }) => (
-  <div style={styles} data-testid="default-today-line" />
-)
+const defaultRenderer = createDefaultRenderer('default-today-line')
 
 /** Marker that is placed based on current date.  This component updates itself on
  * a set interval, dictated by the 'interval' prop.
